+++ src/lib/hamibotApi.ts
@@ -1,12 +1,16 @@
 import { extname } from 'path';
-import * as request from "./request";
 
-export namespace Robot {
+import * as request from "./request";
+import { validRobotId, validScriptId } from './valid';
+
+export class Robot {
+    private constructor() { };
+
     /**
      * @description: 获取机器人列表。
      * @return {Promise<RobotList>} 机器人列表。
      */
-    export async function getRobotList(): Promise<RobotList> {
+    async getRobotList(): Promise<RobotList> {
         return await request.get<RobotList>('/v1/robots');
     }
 
@@ -15,7 +19,9 @@ export namespace Robot {
      * @param {string} robotId 机器人 ID 。
      * @return {Promise<RobotList>} 机器人的详细信息。
      */
-    export async function getRobotById(robotId: string): Promise<RobotItem> {
+    async getRobotById(robotId: string): Promise<RobotItem> {
+        // 校验机器人 ID 格式
+        validRobotId(robotId);
         return await request.get<RobotItem>(`/v1/robots/${robotId}`);
     }
 
@@ -23,7 +29,9 @@ export namespace Robot {
      * @description: 停止指定机器人正在运行的所有脚本。
      * @param {string} robotId 机器人 ID 。
      */
-    export async function stopRobotById(robotId: string): Promise<void> {
+    async stopRobotById(robotId: string): Promise<void> {
+        // 校验机器人 ID 格式
+        validRobotId(robotId);
         await request.put(`/v1/robots/${robotId}/stop`);
     }
 
@@ -34,62 +42,66 @@ export namespace Robot {
      * - `title` - 推送标题，最大长度为 128 。
      * - `text` - 推送内容，最大长度为 512 。
      */
-    export async function sendMessage(robotId: string, msg: PostMessage): Promise<void> {
+    async sendMessage(robotId: string, msg: PostMessage): Promise<void> {
+        // 校验机器人 ID 格式
+        validRobotId(robotId);
         await request.post(`/v1/robots/${robotId}/messages`, msg);
-    }
-
-    export interface RobotList {
-        count: number;
-        items: RobotItem[];
-    }
-
-    export interface RobotItem {
-        /**
-         * @description: 机器人 ID 。24 位 16 进制字符串。
-         * @example 'a51d237e9af41ecc021c9ff6'
-         */
-        _id: string;
-
-        /**
-         * @description: 当前机器人是否在线。
-         */
-        online: boolean;
-
-        /**
-         * @description: 机器人标签。
-         */
-        tags: string[];
-
-        /**
-         * @description: 机器人名称。
-         */
-        name: string;
-
-        /**
-         * @description: 机器人生产厂家。
-         * @example 'Xiaomi'
-         */
-        brand: string;
-
-        /**
-         * @description: 机器人型号。
-         * @example 'MI 11'
-         */
-        model: string;
-    }
-
-    interface PostMessage {
-        title: string;
-        text: string;
     }
 }
 
-export namespace Script {
+export interface RobotList {
+    count: number;
+    items: RobotItem[];
+}
+
+export interface RobotItem {
+    /**
+     * @description: 机器人 ID 。24 位 16 进制字符串。
+     * @example 'a51d237e9af41ecc021c9ff6'
+     */
+    _id: string;
+
+    /**
+     * @description: 当前机器人是否在线。
+     */
+    online: boolean;
+
+    /**
+     * @description: 机器人标签。
+     */
+    tags: string[];
+
+    /**
+     * @description: 机器人名称。
+     */
+    name: string;
+
+    /**
+     * @description: 机器人生产厂家。
+     * @example 'Xiaomi'
+     */
+    brand: string;
+
+    /**
+     * @description: 机器人型号。
+     * @example 'MI 11'
+     */
+    model: string;
+}
+
+interface PostMessage {
+    title: string;
+    text: string;
+}
+
+export class Script {
+    private constructor() { };
+
     /**
      * @description: 获取正在开发中的脚本列表。
      * @return {Promise<ScriptList>} 开发中的脚本列表。
      */
-    export async function getScriptList(): Promise<ScriptList> {
+    async getScriptList(): Promise<ScriptList> {
         return await request.get<ScriptList>('/v1/devscripts');
     }
 
@@ -98,7 +110,9 @@ export namespace Script {
      * @param {string} scriptId 脚本 ID 。
      * @return {Promise<ScriptItem>} 脚本的详细信息。
      */
-    export async function getScriptById(scriptId: string): Promise<ScriptItem> {
+    async getScriptById(scriptId: string): Promise<ScriptItem> {
+        // 校验脚本 ID 格式
+        validScriptId(scriptId);
         return await request.get<ScriptItem>(`/v1/devscripts/${scriptId}`);
     }
 
@@ -108,7 +122,13 @@ export namespace Script {
      * @param {RobotMark} robots 机器人标记。
      * @param {object} scriptConfig 运行时加载的脚本配置。
      */
-    export async function runScript(scriptId: string, robots: RobotMark[], scriptConfig?: object): Promise<void> {
+    async runScript(scriptId: string, robots: RobotMark[], scriptConfig?: object): Promise<void> {
+        // 校验脚本 ID 格式
+        validScriptId(scriptId);
+
+        // 校验机器人 ID 格式
+        robots.forEach((value) => validRobotId(value._id));
+
         let data: any = { robots: robots };
 
         if (scriptConfig) {
@@ -123,7 +143,13 @@ export namespace Script {
      * @param {string} scriptId 脚本 ID 。
      * @param {RobotMark} robots 机器人标记。
      */
-    export async function stopScript(scriptId: string, robots: RobotMark[]): Promise<void> {
+    async stopScript(scriptId: string, robots: RobotMark[]): Promise<void> {
+        // 校验脚本 ID 格式
+        validScriptId(scriptId);
+
+        // 校验机器人 ID 格式
+        robots.forEach((value) => validRobotId(value._id));
+
         await request.del(`/v1/devscripts/${scriptId}/run`, { robots: robots });
     }
 
@@ -132,7 +158,7 @@ export namespace Script {
      * @param {string} scriptName 脚本名称。
      * @return {Promise<ScriptItem>} 创建脚本的详细信息。
      */
-    export async function createNewScript(scriptName: string): Promise<ScriptItem> {
+    async createNewScript(scriptName: string): Promise<ScriptItem> {
         return await request.post<ScriptItem>(`/v1/devscripts`, { name: scriptName });
     }
 
@@ -140,22 +166,24 @@ export namespace Script {
      * @description: 删除指定脚本。
      * @param {string} scriptId 脚本 ID 。
      */
-    export async function deleteScript(scriptId: string): Promise<void> {
+    async deleteScript(scriptId: string): Promise<void> {
+        // 校验脚本 ID 格式
+        validScriptId(scriptId);
         await request.del(`/v1/devscripts/${scriptId}`);
     }
+}
 
-    interface ScriptList {
-        count: number;
-        items: ScriptItem[];
-    }
+interface ScriptList {
+    count: number;
+    items: ScriptItem[];
+}
 
-    interface ScriptItem {
-        _id: string;
-        name: string;
-    }
+interface ScriptItem {
+    _id: string;
+    name: string;
+}
 
-    interface RobotMark {
-        _id: string;
-        name: string;
-    }
+interface RobotMark {
+    _id: string;
+    name: string;
 }