+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-10-06
+
 ### Added
 
 - 全局设置指令:
@@ -65,6 +67,7 @@
 [语义化版本号]: https://semver.org/spec/v2.0.0.html
 
 <!-- Versions -->
-[unreleased]: https://github.com/batu1579/hamibot-assistant/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/batu1579/hamibot-assistant/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/batu1579/hamibot-assistant/compare/v0.0.1..v0.2.0
 [0.1.0]: https://github.com/batu1579/hamibot-assistant/compare/v0.0.1..v0.1.0
 [0.0.1]: https://github.com/batu1579/hamibot-assistant/releases/tag/v0.0.1